requirements.txt
Flask==1.1.2
pycryptodomex==3.9.7